desktop_admin/requirements.txt
streamlit
pandas
streamlit-aggrid
requests
python-dotenv
